Create check_divisible_by_5
This C program checks whether a given number is divisible by 5. It defines a function ChackDivisible that returns true if the input number is divisible by 5, and false otherwise. The main function takes user input, calls the function, and prints the appropriate message. Note: there are a few typos like "Chack" instead of "Check" and "divisibal" instead of "divisible".